takt: fix-decode-and-split-modules
